Rename 'ByStaticMethod' to 'Dynamic'
